port almost all commands and do a few changes
