+++ package.json
@@ -15,12 +15,12 @@
     "url": "https://github.com/Avdan-OS/Tinan/issues"
   },
   "dependencies": {
-    "discord.js": "^14.0.3",
+    "discord.js": "^14.1.2",
     "dotenv": "^16.0.1",
     "octokit": "^2.0.4"
   },
   "devDependencies": {
-    "@types/node": "^18.6.2",
+    "@types/node": "^18.6.3",
     "ts-node": "^10.9.1",
     "typescript": "^4.7.4"
   }
