+++ porting.txt
@@ -1,0 +1,58 @@
+Commands:
+  1. /loa (Leave of Absence) - Lets a developer set when they will leave, the reason why and when they will come back to developing.
+    Prevents managers from demoting the developer.
+
+  2. /checkloa - Checks every developer's leave of absence. (managers only)
+    there's the [LOA] role already
+
+  3. /blacklist - Blacklists a user from the points system.
+
+  4. /reset - Reset user's points.
+
+  5. /demote - Demotes a user. (manager only) half done
+
+  6. /checkpoints - Checks user's points.
+
+  7. /repolist - Lists all the repositories in the Avdan-OS organisation.
+
+  8. /commits - Lists commits of the user in all repositories. Suggested by StoneMc. 
+
+  9. /link - Lets you link your github account to your discord account with Tinan.
+
+Events:
+  1. Download - When a message has "download" in it, the bot will send the same message Helper-bot is sending.
+  
+  2. Reminder - Reminds a developer that they will be demoted soon.
+
+  3. Fireship joke revival - "This has been ... in 100 seconds" remember when we spammed it with EnigmaBot? It's coming back with Tinan :)
+
+  4. Reactions:
+    - akane   =>  🅰️🇰🇦🇳🇪🐱
+    - bread   =>  🍞🇧🇷🇪🇦🇩👍
+    - forgor  =>  💀
+    - honk    =>  🦆🇭🇴🇳🇰👍
+    - oink    =>  🐷
+
+if you're gonna rewrite it, then I don't know what will happen and how will we do it
+it's like i don't know how the hell will we write another command handler or just yoink it from here
+same goes to event handler but you dont like it :skull: so it wont be included there maybe
+
+Bot
+  .command("loa")
+  .handler(loaHandler);
+
+for (const {name, handler} of commandHandlers) {
+  Bot
+    .command(name)
+    .handler(handler);
+}
+
+for (const {regex, reaction} of reactions) {
+  Bot
+    .message(regex)
+    .react(reaction);
+}
+
+Bot
+  .message(/akane/i)
+  .react("🅰️🇰🇦🇳🇪🐱");
